+++ CHANGELOG.md
@@ -1,1 +1,3 @@
+  * Always use varnish if its available.
+  * Add expire helper for setting cache control headers.
   * Fixes issue with urls without paths.